Try renaming the folder vs a capitalization change to get git to commit the change.
