+++ src/shared/ListRepo/ReposTable/NoReposBlock.spec.js
@@ -1,7 +1,7 @@
 import { render, screen } from 'custom-testing-library'
 import NoReposBlock from './NoReposBlock.js'
 import { MemoryRouter, Route } from 'react-router-dom'
-import { ActiveContext } from 'shared/contexts'
+import { ActiveContext } from 'shared/context'
 
 describe('NoReposBlock', () => {
   function setup({ owner, active }) {
