+++ src/shared/ListRepo/ReposTable/ReposTable.js
@@ -12,7 +12,7 @@ import NoReposBlock from './NoReposBlock'
 
 import { useFlags } from 'shared/featureFlags'
 import { useContext } from 'react'
-import { ActiveContext } from 'shared/contexts'
+import { ActiveContext } from 'shared/context'
 
 const tableActive = [
   {
